requirements_livekit.txt
# LiveKit Integration Requirements
livekit-agents>=0.8.0
livekit-api>=0.5.0
livekit-rtc>=0.9.0

# LiveKit plugins (uncomment as needed)
livekit-plugins-openai
# livekit-plugins-deepgram
# livekit-plugins-elevenlabs

# Environment management
python-dotenv>=1.0.0

# Additional async support
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Audio processing
soundfile>=0.12.0
resampy>=0.4.0